es6/es7 changes

+++ android/app/src/main/java/com/microsoft/codepush/react/CodePushPackage.java
@@ -225,9 +225,11 @@ public class CodePushPackage {
         }
     }
 
-    public void clearUpdates() {
-        File statusFile = new File(getStatusFilePath());
-        statusFile.delete();
-        CodePushUtils.deleteDirectoryAtPath(getCodePushPath());
+    public void clearTestUpdates() {
+        if (CodePush.isUsingTestConfiguration()) {
+            File statusFile = new File(getStatusFilePath());
+            statusFile.delete();
+            CodePushUtils.deleteDirectoryAtPath(getCodePushPath());
+        }
     }
 }
